+++ src/Module.php
@@ -14,6 +14,7 @@ use luya\admin\importers\AuthImporter;
 use luya\admin\importers\FilterImporter;
 use luya\admin\importers\PropertyImporter;
 use luya\admin\filesystem\LocalFileSystem;
+use luya\admin\base\ReloadButton;
 
 /**
  * Admin Module.
@@ -30,6 +31,9 @@ use luya\admin\filesystem\LocalFileSystem;
  * ]
  * ```
  *
+ * @property array $reloadButtons Take a look at {{luya\admin\Module::setReloadButtons()}}.
+ * @property array $jsTranslations Take a look at {{luya\admin\Module::setJsTranslations()}}.
+ * 
  * @author Basil Suter <basil@nadar.io>
  * @since 1.0.0
  */
@@ -350,6 +354,59 @@ final class Module extends \luya\admin\base\Module implements CoreModuleInterfac
         $this->_jsTranslations = $translations;
     }
     
+    private $_reloadButtons = [];
+
+    /**
+     * Set an array of relaod buttons with a callback function to run on click.
+     * 
+     * Every array item needs at least:
+     * 
+     * + label: The label which is displayed in the mnu
+     * + icon: A material icon value from https://material.io/tools/icons/
+     * + callback: A php callable function which is executed when clicking the button.
+     * 
+     * ```php
+     * 'reloadButtons' => [
+     *     ['label' => 'Clear Frontpage Cache', 'icon' => 'clear', 'callback' => function($button) {
+     *         (new \Curl\Curl())->get('https://luya/clear/this/cache');
+     *     }]
+     * ]
+     * ```
+     * 
+     * The first paramter of the callback function is the ReloadButton object itself, this allwos you to
+     * change the response message.
+     * 
+     * ```php
+     * 'callback' => function(\luya\admin\base\ReloadButton $button) {
+     *     // do something
+     *     // ...
+     * 
+     *     // change response (success) message.
+     *     $button->response = 'Running this button was a full success!';
+     * }
+     * ```
+     * 
+     * @param array $buttons
+     * @since 2.0.0
+     */
+    public function setReloadButtons(array $buttons)
+    {
+        foreach ($buttons as $buttonConfig) {
+            $this->_reloadButtons[] = new ReloadButton($buttonConfig);
+        }
+    }
+    
+    /**
+     * Return array with {{luya\admin\base\ReloadButton}} objects
+     *
+     * @return array
+     * @since 2.0.0
+     */
+    public function getReloadButtons()
+    {
+        return $this->_reloadButtons;
+    }
+
     /**
      * Get the admin module interface menu.
      *