+++ src/messages/nl/admin.php
@@ -390,4 +390,7 @@ return [
     'login_browsehappy' => 'You are using an <strong>outdated</strong> browser. Please <a href="https://browsehappy.com/" target="_blank">upgrade your browser</a> to improve your experience and security.',
     'crud_notification_enable' => 'Enable Notification',
     'crud_notification_disable' => 'Disable Notification',
+    'admin_button_execute' => '"{label}" executed',
+    'layout_btn_reload_admin' => 'Reload&nbsp;Admin',
+    'layout_btn_reload_cache' => 'Clear&nbsp;Cache',
 ];